3/19/2024 upload
I added the correct typefaces to where they go. I resized my portfolio images. I added the text to the info popups. I worked on formatting the form section. 

I noticed that the clickbox for the logo button is not quite properly positioned. I still need to position sections in the correct place.